TTF fix
